Disable the Sign in button if the form is invalid
The sign in form is invalid if the email doesn't match the regexp or is
empty, or if the password is empty.

This change doesn't set a min or max length requirement for email or
password, as the sign up form does. This allows the length requirements
to change on the server, but still, permit existing users to sign in.