Don't display Sign Out page after signing in
If the user signs out using the Sign Out page they are returned to the
Sign In page, if the user signs in they should not be returned to the
Sign Out page.

This bug was a regression introduced when the "No" button was added to
the Sign Out page.