Rewrite error message lookup
The new implementation will attempt to read any `message` value in a
JSON object returned with a 400 status. If no message is found a table
lookup is performed, the table contains strings for known error codes.
For non-400 status error the implementation falls back to
QNetworkReply::errorString().

Also:
- Removed BaseResponse::httpStatusReason(). Its return value is now set
as the BaseResponse::errorMessage(), so that should be used instead.
- Changed BaseResponse::getHttpStatusReason() to return an empty string
if the HTTP status code is not handled, previously to delegated to
BaseResponse::errorMessage() which would have been equal to
QNetworkReply::errorString().
- Renamed BaseResponse::setError() -> BaseResponse::setErrorMessage()