Avoid displaying chart data from old requests
The user can change the charts configuration (i.e data-range) which
causes a new request for data to be sent. If the user changes the
configuration before the previous request was finished and the new
request finished first the old request data will be displayed.

This change avoids this bug by ignoring old requests.