Move price prediction button to chart toolbar
The price prediction button's text was originally going to display the
actual price prediction statement, since that is not possible the text
is now a string literal that looks out of place by itself below the
chart. This commit moves the button to the chart toolbar which looks a
little nicer.