Fix swipe hidden actions on Portfolio, Watchlist
The actions are "Close Position" on the Portfolio page, and "Delete"
watchlist item on the Watchlist page.

This was broken in 69c94a9b2face1e2ff7543deb8fc4acacfad129f "Fix pressed
state color of swipe hidden actions".

Using a MouseArea instead of SwipeDelegate.onClicked is a workaround for
Android. This commit fixes that workaround and still correctly shows the
pressed state.